chg: refactor(tags): centralize metadata handling and title generation
Extracts metadata reading logic for MP3 and FLAC files from `tags.go` into a new, dedicated `metadata_reader.go` file. This improves code organization and modularity.

A new `song.GetTitle()` method is introduced to provide a single source for generating track titles, which now includes version information. The FLAC and ID3v2 tagging processes have been updated to use this method.

As a result, filename generation in the migration logic is simplified to use the track number directly, removing the need for parsing and reformatting.
