+++ internal/tags/id3v2.go
@@ -37,7 +37,7 @@ func (t *id3v2Tagger) addTags(resource deezer.Resource, song *deezer.Song, cover
 		t.addTag("TRCK", song.TrackNumber)
 	}
 
-	t.addTag("TIT2", song.Title)
+	t.addTag("TIT2", song.GetTitle())
 	t.addTag("TPE1", strings.Join(song.Contributors.MainArtists, ", "))
 	t.addTag("TCOM", strings.Join(song.Contributors.Composers, ", "))
 	t.addTag("TEXT", strings.Join(song.Contributors.Authors, ", "))
